// admin/src/types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      exercises: {
        Row: {
          aliases: string[]
          benefits: string | null
          can_be_activation: boolean
          category: string | null
          created_at: string | null
          difficulty: string | null
          id: string
          media_url: string | null
          movement_pattern: string | null
          name: string
          name_eng: string | null
          primary_muscles: string[]
          reps_range: string | null
          risks: string | null
          search_text: string | null
          secondary_muscles: string[] | null
          settings: string | null
          short_name: string | null
          status: string
          tags: string[]
          technique: string | null
          updated_at: string
        }
        Insert: {
          aliases?: string[]
          benefits?: string | null
          can_be_activation?: boolean
          category?: string | null
          created_at?: string | null
          difficulty?: string | null
          id?: string
          media_url?: string | null
          movement_pattern?: string | null
          name: string
          name_eng?: string | null
          primary_muscles: string[]
          reps_range?: string | null
          risks?: string | null
          search_text?: string | null
          secondary_muscles?: string[] | null
          settings?: string | null
          short_name?: string | null
          status?: string
          tags?: string[]
          technique?: string | null
          updated_at?: string
        }
        Update: {
          aliases?: string[]
          benefits?: string | null
          can_be_activation?: boolean
          category?: string | null
          created_at?: string | null
          difficulty?: string | null
          id?: string
          media_url?: string | null
          movement_pattern?: string | null
          name?: string
          name_eng?: string | null
          primary_muscles?: string[]
          reps_range?: string | null
          risks?: string | null
          search_text?: string | null
          secondary_muscles?: string[] | null
          settings?: string | null
          short_name?: string | null
          status?: string
          tags?: string[]
          technique?: string | null
          updated_at?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}